+++ docs/changelog.md
@@ -1,5 +1,8 @@
 # UniK Changelog
 
+## Thu, Jul 28 2016
+* UniK now supports running [IncludeOS](http://www.includeos.org/) Unikernels (for C++). Currently only the Virtualbox provider can run IncludeOS unikernels.
+
 ## Fri, Jun 17 2016
 * UniK's Containers are now (automatically) versioned by the first 16 characters of their SHA256 checksum.
 * Changed OSv / Java support: